Implement screen to create backup mesage and its signature
Summary:
This differential implements screen in registration flow to create backup message and its signature. It is consistent with the following figma design:
https://www.figma.com/file/a1nkbWgbgjRlrOY9LVurTz/Comm-%2F-Desktop-app?type=design&node-id=11617-293984&mode=design&t=kZlddBvG5d3ZbRFk-0. The screen is placed as the last step
in registration flow. As of this differential generated siwe message and signature is logged. Next diffs will implement relevant JSI to put in the SQLite.

Test Plan: Execute registration flow. Ensure that we go through backup key generation step. Ensure that message and signature is logged to the console.

Reviewers: tomek, inka, atul, ginsu, ashoat

Reviewed By: atul, ginsu, ashoat

Subscribers: ginsu, ashoat

Differential Revision: https://phab.comm.dev/D11636