[lib][web][native] Pass preRequestUserState into invalidSessionRecovery
Summary:
issue: [[https://linear.app/comm/issue/ENG-6126/pass-prerequestuserstate-into-invalidsessionrecovery-for | ENG-6126]]
We want to cross reference the user info from before the action was called with the current user info, to check if the same user is still logged in. Otherwise we could end up overriding a new users session with a recovery that was startd for the previous user.
Test Plan:
Tested on both platforms that normal logins and loguts work as expected
Tested on native that if a `COOKIE_INVALIDATION_RESOLUTION_ATTEMPT` login with a different user id is dispatched, then the action is disregarded

Reviewers:
tomek, ginsu
Subscribers: