Extract shareable logic from CreateSIWEBackupMessage component
Summary:
This differential refactors `CreateSIWEBackupMessage` component so that most of its code and logic can be resused. It is necessary since next diff will
introduce new component that renders the same UI to already logged in SIWE users to create backup message if it is not in the database.

Test Plan:
1. Flow
2. Test that already introduced functionality wasn't broken.

Reviewers: tomek, atul, inka

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11644