[web] Reassign the notifs session
Summary:
If a cookie changes move the session from one place to a new one. If a cookie becomes null, we don't want to reassign - it might mean that e.g. a user logged out.

https://linear.app/comm/issue/ENG-7689/reassign-the-session

Depends on D11648

Test Plan:
Open the web app, login, send a message from another client - a notif should be displayed (without this stack we can only see a notif with a warning message). Also, verify that the reassignment actually occurs by adding some console logs.

Opened two web tabs, and tried to log in on both of them. Sent a message from another client and confirmed that a notif was displayed.

Reviewers: kamil, marcin, varun

Reviewed By: marcin

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11649