[native][android] Move database and MMKV initialization to CommHybrid
Summary:
This diff prevents the Android app from getting into a state that causes a crash loop, described in [ENG-7696](https://linear.app/comm/issue/ENG-7696/commsecurestore-crash-during-commmmkvinitialize). It however does not resolve the crash loop if it already occurring; that is addressed in the following diff.

`CommHybrid` is perhaps is not the right place for these calls, since it's meant to be a proxy to the C++ version of `CommHybrid`. Open to other suggestions.

Test Plan: Deleted the app off the simulator and redeployed, and the problem was gone (app stopped crash looping after being closed/reopened)

Reviewers: marcin, tomek

Subscribers: