[native] Wait for authoritative keyserver login before setting avatar in registration flow
Summary:
When this logic was first built, `currentUserInfo` would appear after the authoritative keyserver auth occurred.

Now it appears after identity auth, so it's no longer the right thing to check here. Instead, we should check `isLoggedInToKeyserver(authoritativeKeyserverID)`.

Test Plan: Make sure that I can register an account with a custom avatar, and that when the account is created the custom avatar appears

Reviewers: inka, tomek, marcin

Reviewed By: marcin

Differential Revision: https://phab.comm.dev/D11693