[native] Purge the store after clearing the DB
Summary: `clearSensitiveData` clears the token, which causes an event to be emitted, which causes `AccessTokenHandler` to dispatch a new action. This action is dispatched after the store is purged and causes a new state to be persisted. The solution is to purge after the clearing of the DB is finished.

https://linear.app/comm/issue/ENG-7788/state-difference-after-state-wipe

Test Plan: Wipe the state, start the app - there should be no state diff.

Reviewers: kamil, inka