[lib][native][web] Move useLegacyAshoatKeyserverCall to lib/keyserver-conn/legacy-keyserver-call.js
Summary:
`lib/utils/action-utils.js` used to be a huge, unwieldy file. Now all that's left is `useLegacyAshoatKeyserverCall`, so the file can be renamed.

Additionally, I'm moving it into `lib/keyserver-conn`. Given the name and function, that seems like a better place, especially since `useKeyserverCall` was moved there as well in the prior diff.

Depends on D11699

Test Plan: Flow

Reviewers: tomek, inka

Subscribers: