[web] Remove authoritativeKeyserverID check from getInitialReduxState
Summary:
issue: [[https://linear.app/comm/issue/ENG-6008/revert-ashoatkeyserverid-check-in-getinitialreduxstate#comment-2e0a433f | ENG-6008]]
This was a hack needed for testing adding keyserver back when we didn't have multikeyserver setup. As discussed in the issue, it is no longer needed

Test Plan:
Have been using multikeyserver setup with this change as part of the patch diff D11141
I was able to add keyservers and communicate with them. Calling this endpoint worked correctly.

Reviewers: ginsu

Reviewed By: ginsu

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11701