[keyserver] introduce barebones createOrUpdateFarcasterChannelTag
Summary: This diff introduces the barebones `createOrUpdateFarcasterChannelTag` function. For the scope of this project we just want to upload a mapping of farcaster channel => comm community to the blob service, with the intention that all the other necessary parts of this function like (permissions, inserting the blob holder in the keyserver will be addressed as a follow up project that I will start in the next week)

Test Plan: In my local stack I was able to use `createOrUpdateFarcasterChannelTag` to upload the farcaster channel => comm community blob and got the expected response back.

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11772