[keyserver] Use keyserverID instead of userID when creating invite link blobs
Summary: For each link we create a blob that contains a keyserver ID that is later used as a part of the override on the client side. In a single-keyserver world, this value should be equal to 256 because this is the ID a client uses as a keyserver ID. Instead of using the correct value, we were using the ID of a user, taken from Identity - so it was a multi-keyserver keyserver ID.

This issue wasn't affecting prod because the ID of the keyserver on prod Identity is equal to 256.

https://linear.app/comm/issue/ENG-7789/invite-links-in-single-keyserver-world-fail

Test Plan: Tested in creating a new link and accepting it works in both single- and multi-keyserver worlds.

Reviewers: inka