[keyserver] introduce deleteFarcasterChannelTag
Summary: This diff introduces the barebones `deleteFarcasterChannelTag` function. For the scope of this project we just want this function to delete the farcaster channel tag blob from the blob service for now, with the intention that all the other necessary parts for this function (like permissions, updating the row in the community table on the keyserver to remove the blob holder will be added to this function as a follow up project that I will start shortly)

Test Plan: In my local stack I was able to use `deleteFarcasterChannelTag` to delete the farcaster channel => comm community blob. I also confirmed that when the `GATE_TAG_FARCASTER_CHANNEL` gate is set to `true` the function throws a server error

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11813