[web][native][keyserver] Flip the switch to enable `specialRole` field
Summary:
This diff will need to be landed at the same time as a release/`codeVersion` bump so things don't break in developer's local environment.

Will send message in Releases channel when ready.

---

Depends on D11188

Test Plan: Have a local "do not land" diff that does this locally and things work as expected.

Reviewers: ashoat, ginsu, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D11204