[SQLite/lib] implement shared SQLite interface
Summary:
Similar to D11111 but for SQLite.
We want to use API to received messages form `lib/`. We could unify more of existing code to fit into this interface but this will require some more changes to make types consistent.

Depends on D11725

Test Plan:
Test this interface from `/lib`.

Reviewers: tomek, marcin