[lib] Fix error when deleting a keyserver
Summary:
issue: [[https://linear.app/comm/issue/ENG-7174/deleting-a-keyserver-results-in-errors | ENG-7174]]
Deleting a keyserver would result in the app crashing, because `updateConnectionStatusActionType` would be dispatched for the removed keyserver.
We should just ignore this action. This happens because of unfortunate ordering - the keyserver is removed, and before the socket is unmounted the action gets dispatched

Test Plan: Tested that after this the app no longer crashes after removing a keyserver from the keyserver list.

Reviewers: tomek, kamil

Reviewed By: tomek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11736