[lib] Don't set dataLoaded on authoritative keyserver auth
Summary:
For the registration flow, authoritative keyserver auth concludes after step 2, but there is a step 3. We should wait on step 3 before setting `dataLoaded`, and in fact we already have a dispatch of `setDataLoadedActionType` in `useRegistrationServerCall`.

Now that authoritative keyserver auth doesn't set `dataLoaded`, I had to add a dispatch of `setDataLoadedActionType` to `useLogIn` in order to handle it manually.

Depends on D11768

Test Plan:
This whole stack was tested with the following steps:

1. On native, attempting to log in with an ETH account that hasn’t been registered yet
2. On native, attemping to register with an ETH account that has already been registered
3. On native, log in with an ETH account that has already been registered
4. On native, register with an ETH account that hasn’t already been registered
5. On native, register with a password account
6. On native, log in with a password account
7. On web, attempting to log in with an ETH account that hasn’t been registered yet
8. On web, log in with an ETH account that has already been registered
9. On web, log in with a password account

Reviewers: tomek, inka