[keyserver] Print exception that causes failed_to_retrieve_inbound_keys
Summary: This is useful for debugging. I needed it in order to diagnose [ENG-7936](https://linear.app/comm/issue/ENG-7936/failedprecondition-genericfailure-when-querying-staging-identity-for), so I figured it would be good to add to the codebase in case we need to diagnose a similar issue in production in the future.

Test Plan: Confirm that relevant log was printed (see the Linear issue for the log)

Reviewers: will, varun, bartek

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D11770