[web] Fix missing relationship buttons
Summary:
issue: [[https://linear.app/comm/issue/ENG-7486/relationship-buttons-missing-on-web | ENG-7486]]
We might not have user infos in our UserStore, if we don't have any relationship with them. But we can extract them from the pending thread infos. `useRelationshipPrompt` takes `pendingPersonalThreadUserInfo` as an argument, and uses it if the user infos are indeed not present in the store.

Test Plan: Tested that the relationship buttons are present and work correctly

Reviewers: tomek, kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11760