[keyserver] Fix sending a user cookie when olm session creation failed
Summary:
issue: [[https://linear.app/comm/issue/ENG-7247/its-possible-to-be-logged-in-and-not-have-a-csat | ENG-7247]]
It used to be possible to have a user cookie even though the notif olm session creation failed. This was because the cookie was created before the session, so even though an error was thrown, the user cookie would be sent with it.

We want to create the user cookie once both olm sessions are created. The only thing I really did here is moving code around so that the cookie creation is called after both session are created

Test Plan: Tested that if an error if thrown from `persistOlmNotifSessionPromise`, the user doesn't get a user cookie and is no longer logged in after refreshing the app

Reviewers: varun, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11780