[lib][native][web] Display a message when accepting an invite link times out
Summary:
In order to support a timeout correctly on both platforms we had to change how we store the status on native.
Added a new status which represents a timeout and is handled in the same way as an invalid link - the only difference is a displayed message.

https://linear.app/comm/issue/ENG-7924/improve-invite-link-join-timeout-handling

Web timeout
{F1667476}

Web invalid
{F1667473}

Web success
{F1667472}

Native timeout
{F1667477}

Native invalid
{F1667474}

Native success
{F1667475}

Test Plan:
Tested on native and web:
1. Added a timeout on the keyserver and checked if the timeout message appears
2. Thrown an exception on the keyserver and checked if the invalid link message appears
3. Checked if it is possible to join a community

Reviewers: inka, ashoat

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D11812