[keyserver] introduce barebones createOrUpdateFarcasterChannelTag
Summary:
This diff introduces the barebones `createOrUpdateFarcasterChannelTag` function. For the scope of this project we just want this function to upload a mapping of farcaster channel => comm community to the blob service for now, with the intention that all the other necessary parts for this function like (permissions, inserting the blob holder in the keyserver will be addressed and added to this function as a [[ https://linear.app/comm/issue/ENG-7713/introduce-communities-table-to-keyserver | follow up project ]] that I will start shortly)

Linear task: https://linear.app/comm/issue/ENG-7575/introduce-uploadfarcasterchannelblob-function and https://linear.app/comm/issue/ENG-7581/introduce-getfarcasterchannelblob-function

Test Plan:
In my local stack I was able to use `createOrUpdateFarcasterChannelTag` to upload the farcaster channel => comm community blob and got the expected response back.

I also confirmed that when I tried to reuse the same farcaster channel tag I would trigger the `already_in_use` error which is expected

Reviewers: bartek, atul, tomek

Reviewed By: tomek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11772