[lib] introduce fetchFarcasterChannelByName to neynar client class
Summary: `fetchFarcasterChannelByName` is a function that given a farcaster channel name, will retreive the info for that channel. The reason for this function is that we want to validate the user input given to us from the `TagUnfollowedFarcasterChannel` screen and confirm that the user inputted a valid channel name.

Test Plan: Used the neynar docs to confirm that `fetchFarcasterChannelByName` was getting the expected/same response as the docs.

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11865