[lib] Propagate exceptions in keyserverAuth to callers
Summary:
I had made this change initially in D11500, but I [forget why](https://phab.comm.dev/D11500?id=38613#inline-69293) so I undid it.

The reason for the change is that `dispatchActionPromise` will not reject if the promise passed to it rejects.

We want failures in `keyserverAuth` to propagate to callers. At all of its callsites, `keyserverAuth` calls are wrapped in a try-catch where we need to do something on failure:

1. In `KeyserverConnectionHandler` [here](https://github.com/CommE2E/comm/blob/4ba9c88fe6042bba34851ace0985dd7c19b62c09/lib/keyserver-conn/keyserver-connection-handler.js#L99-L106)
2. In `useLogIn` [here](https://github.com/CommE2E/comm/blob/4ba9c88fe6042bba34851ace0985dd7c19b62c09/lib/hooks/login-hooks.js#L115-L121)
3. In `useRegistrationServerCall` [here](https://github.com/CommE2E/comm/blob/4ba9c88fe6042bba34851ace0985dd7c19b62c09/native/account/registration/registration-server-call.js#L332-L339)

Test Plan:
1. Force keyserver auth to fail with [this patch](https://gist.github.com/Ashoat/0029722313ffb21d619fa28af786f1ae)
2. Make sure `LOG_OUT_STARTED` and `LOG_OUT_SUCCESS` are dispatched when keyserver auth fails

Reviewers: tomek, inka

Reviewed By: tomek

Differential Revision: https://phab.comm.dev/D11837