[lib] Display a dev prefix for dev invite links
Summary:
This is an approximate solution. It won't work if someone decides to e.g. server their keyserver from a different port / url. But still, it is an improvement, which will work for most of the cases, will simplify the testing, and will significantly decrease the number of false alarms regarding links not working.

https://linear.app/comm/issue/ENG-7319/display-dev-prefix-on-dev

Test Plan: Check if a link is displayed correctly on dev web.

Reviewers: inka