[lib] add device list type to `AuxUserInfo`
Summary:
[ENG-7394](https://linear.app/comm/issue/ENG-7394/persist-peer-list).

This is simple diff to add device list type to where it'll be store and accessed.
With this type this works fine without migration, we could also update the type to `deviceList: ?RawDeviceList` but then we'll need a migration that reads the data from DB, converts it to objects, adds `deviceList: null` and saves it in DB to make it safe - don't have a strong opinion about it.

We want this field to be nullable, this will indicate that we have a new user, and we need to fetch device list for it from Identity, that will be implemented in [ENG-7824](https://linear.app/comm/issue/ENG-7824/update-peer-list-when-updating-user-store). (cc. @inka)

Test Plan: FLow

Reviewers: tomek, will, bartek!

Subscribers: ashoat, inka

Differential Revision: https://phab.comm.dev/D11840