[web] display modal explaining why user was logged out
Summary:
when web clients receive the first code version with usingCommServicesAccessToken set to true, we will need to log them out so that they can register/log in to the identity service and get a CSAT. when we log them out, we should display a modal explaining what has happened.

Depends on D11705

Test Plan:
set the usingCSAT flag to true and refreshed the app. confirmed that modal was displayed and the link worked

{F1683797}

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D11844