[lib][native] Automatically retry keyserverAuth during login/registration
Summary:
This addresses [ENG-7661](https://linear.app/comm/issue/ENG-7661/retry-keyserver-auth-on-olm-session-creation-failure-with-bad-message).

When `keyserverAuth` is called during login or registration, if it fails we want to immediately try again. This is to guard against a situation where we are vended an OTK that is invalid. While we think we've addressed all scenarios where OTKs can be invalid, it's possible that we will introduce some new scenarios in the future, and we've had sporadic reports of this (eg. [ENG-7927](https://linear.app/comm/issue/ENG-7927/olmbad-message-key-id)).

On the other hand, when `keyserverAuth` is called by `KeyserverConnectionHandler` for auth with secondary keyservers, or for keyserver session recovery, we'll continue only trying once. For secondary keyserver auth we'll end up retrying after `AUTH_RETRY_DELAY_MS` anyways. And for keyserver session recovery, the recovery attempt is arguably already the second try.

Note – would normally put @kamil on this review since we came up with the idea together in a 1:1, but since he's out for the rest of this week, I'm asking @varun to review.

Depends on D11847

Test Plan:
1. Force keyserver auth to fail with [this patch](https://gist.github.com/Ashoat/0029722313ffb21d619fa28af786f1ae)
2. Add some logs that get printed inside keyserver auth
3. Test new registration flow and confirm that the logs get printed twice, but otherwise that the behavior is the same

Reviewers: varun

Subscribers: varun, inka, tomek, kamil

Differential Revision: https://phab.comm.dev/D11848