[native] implement onPressTagChannel for TagUnfollowedFarcasterChannel
Summary: This diff implements the logic for `onPressTagChannel` in `TagUnfollowedFarcasterChannel`. The important thing about this method is that Before dispatching our action promise, we want to validate the input through `fetchFarcasterChannelByName`

Test Plan: Confirmed that `fetchFarcasterChannelByName` correctly validates the input and `createOrUpdateFarcasterChannelTag` action is dispatched when the input is valid. Otherwise confirmed that I was consoling an error log

Reviewers:

Subscribers: