[lib] Shim unsupported messages for `web`/`desktop`
Summary:
We currently aren't shimming messages for `web`/`desktop`, probably because we weren't previously persisting things on those platforms?

However, now that we are persisting messages we'll want to shim/unshim messages. Specically we want to shim/unshim the new `UPDATE_RELATIONSHIP` message type.

Test Plan: WARNING: I'm still reading through codebase to make sure that this change is safe and can be made as-is. I also haven't yet tested this function with `UPDATE_RELATIONSHIP` messages to make sure it behaves as expected. Just putting this up to reference in comment on D11707

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D11712