[lib] Show other user FID in FARCASTER_MUTUAL robotext
Summary:
To look up their Farcaster username we'll need to integrate Neynar into keyserver/web, but we can display the FID now.

I decided to start with the FID so I could test the general machinery and ended up finding some issues, which are addressed in D11870 and D11871.

Test Plan:
| creator view | recipient view |
| {F1697690} | {F1697691} |

Reviewers: ginsu, atul

Reviewed By: ginsu

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D11872