[lib] Show other user FID in FARCASTER_MUTUAL robotext
Summary:
To look up their Farcaster username we'll need to integrate Neynar into keyserver/web, but we can display the FID now.

I decided to start with the FID so I could test the general machinery and ended up finding some issues, which are addressed in the preceding diffs.

Test Plan: [screenshot]

Reviewers: ginsu, atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D11872